Update index.js to parse the PDF
The program separates the text in a PDF into main points and normal text information. The main points are recognized if they are bolded and is larger than the normal text font size. The main points and it's respective text information is saved into a JSON called "textbook". There are still some issues where main points are not recognized properly and the string saved is not accurate to the actual textbook.